libdrgn: orc_info: use unsigned int instead of size_t for num_entries
It's unrealistic for there to be more than 4 billion ORC entries. Switch
to an unsigned int. The main benefit is that the indices array that we
use to sort the parallel arrays of entries and pc_offsets becomes half
the size, which also makes parsing ORC about 10% faster (down from ~5 ms
to ~4.5 ms for the Fedora vmlinux on my laptop).

+++ libdrgn/orc_info.h
@@ -59,7 +59,7 @@ struct drgn_module_orc_info {
 	 */
 	struct drgn_orc_entry *entries;
 	/** Number of ORC unwinder entries. */
-	size_t num_entries;
+	unsigned int num_entries;
 };
 
 void drgn_module_orc_info_deinit(struct drgn_module *module);
